Basic chatbot & training of LSTM model implemented
Using Ollama a basic chatbot assistant was implemented. It currently sources reference material for postgreSQL database and forms answer based on sentiment analysis of current market.

Using the sentiment data added to the dataset, an LSTM model was trained to predict the daily sentiment based on economic factors such as GDP, inflation rate, and property prices.